[Author: cprince]
Remove problematic "Entries.Log" and "Entries.log" files.
Filename case sensitivity is not supported on Windows (P4 or SVN).

Thse files are only in /art/CVS, and the directory also has Entries
(which is used everywhere else in the sqlite tree), so it seems
correct to delete the ".Log" and ".log" variations.

PRESUBMIT=passed
TBR=shess
OCL=4825230
SCL=4825230