[Author: cprince]
Settle on "allowCrossOrigin" for API name.

Background: Aaron always liked this name, and I favor it again, now that
we've realized the behavior is somewhat specialized (it really activates
the factory, only in a child, only when created cross-origin).  This is
hard to convey in a concise name, and the details may change later.

R=michaeln
CC=google-gears-eng@googlegroups.com
DELTA=100  (5 added, 0 deleted, 95 changed)
OCL=5268817
SCL=5270433