[Author: aa]
Change the functionality of workerPool.onerror. Instead of having to be set
on parent workers and reporting all errors from all children, it should be
set on child workers and it reports errors that happen in that child.

Also, to be like window.onerror, workerpool.onerror can return a boolean
true to indicate that the error has been handled and should not be reported
to the parent.

Otherwise, the error is rethrown globally in the parent worker.

Change includes updated unit tests and documentation.

R=cprince
CC=google-gears-eng@googlegroups.com
DELTA=364  (235 added, 51 deleted, 78 changed)
OCL=5423098
SCL=5431959