[Author: aa]
Make the tests run serially instead of parallel. This was necessary to
prevent some edge cases during tests where two tests can collide on the
same URL.

Now, file_test_results is far over complicated for what it does. It could
be a simple state machine iterating over all the tests in the config.

In the interest of expediency, I'm making the minimum change required to
make the tests run serially. We can refactor the code later.

There are also a few unrelated changes in assert.js, because I started on
this in the middle of porting the other tests. Sorry about that.

R=ace
CC=gears-internal
DELTA=255  (155 added, 51 deleted, 49 changed)
OCL=5631676
SCL=5632490