[Author: aa]
Some things I finished on various airplanes. Port the workerpool and
httprequest tests and fix some issues in the localserver tests.

Note: These tests still fail ocassionally because of race conditions
between asynchornous tests completing and test runner callbacks. In
retrospect, the only way to make this both reliable and reasonably fast
is to let the test control when the test is over, not the test runner.

I have an idea of how to do this that does not require lots of functional
programming and will work on it on the next plane :).

Otherwise, every test is working except:
-

R=ace
CC=gears-internal
DELTA=935  (881 added, 5 deleted, 49 changed)
OCL=5660750
SCL=5706752