[Author: aa]
Return S_OK instead of E_FAIL to OnScriptError(). This was causing a bug
where IE would keep trying OnScriptError() for each eval() call on the
stack, resulting in multiple calls to worker.onerror.

We used to return E_FAIL here so that we could detect failure of the
initial script eval in worker setup. But we have since changed that so
that we consider workers valid even if initial script evaluation fails.

R=cprince
CC=google-gears-eng@googlegroups.com
DELTA=27  (10 added, 1 deleted, 16 changed)
OCL=5710447
SCL=5725012