[Author: aa]
Fix a few bugs around destroying nested workerpools and workerpools with
odd lifetimes. The old unit tests did not have any code like this, so I am
not surprised we didn't see it before. I also added a unit test to
explicitly test one of the cases, even though there was already a test that
hit it on accident.

===

Crash with the following code because the nested pool tries to call
JS_RemoveRoot for worker 0 when the corresponding engine is already gone.

   var f = new GearsFactory();
   var wp = f.create('beta.workerpool', '1.1');
   wp.createWorker(
     'var wp = google.gears.factory.create("beta.workerpool", "1.1");' +
     'wp.onmessage = function(){};'
   );

===

   Assert with the following code because the httprequest ends up getting
   deleted from a child worker when it was created on the owning worker.

   function foo() {
     var f = new GearsFactory();
     var wp = f.create('beta.workerpool', '1.1');
     wp.createWorkerFromUrl('anything.js');
   }
   foo();

===

   A related assert to the above where factory_ref which was created on a
   child worker ended up getting deleted from the owning worker.

R=zork
CC=google-gears-eng@googlegroups.com
DELTA=32  (32 added, 0 deleted, 0 changed)
OCL=5729727
SCL=5732451