[Author: mpcomplete]
Initial work on NPAPI port of Gears.

This is very much a work in progress.  There are a number of places where I just aliased BROWSER_NPAPI for BROWSER_IE so that it would use the win32-specific implementations of things.  Eventually these will have to be separated out into OS-specific implementations.

The NPAPI binding layer is still not hooked up yet.  I wanted to check this in so others could start hacking on it.  My plan is to have a set of NPObject-derived classes, one for each gears module, that serve as the bridge between NPAPI and gears.  So for instance, GearsFactoryPlugin is the wrapper around GearsFactory, which will have the implementation.  The names are not set in stone.

R=cprince
CC=gears-internal
DELTA=2815  (2809 added, 2 deleted, 4 changed)
OCL=5743290
SCL=5762472