[Author: michaeln]
Fix a bug in the HttpHandler for IE where a POST could redirect to resource in the
localserver but our request intercept layer would not intercept it. We now run more
requests in 'passthru mode' to detect when such a redirect occurs and to intervene.
Plus a unit test.

R=aa
CC=google-gears-eng@googlegroups.com
DELTA=65  (53 added, 3 deleted, 9 changed)
OCL=5930323
SCL=5931218