[Author: mpcomplete]
Fix a memory leak in JsArray and JsObject for the IE and NPAPI ports.  Each time you call GetElement* or GetProperty* to retrieve string or object token, the token would be leaked.  I added a JsScopedToken to ensure that the right thing is done when the object leaves scope, depending on the platform.

R=zork
CC=google-gears-eng@googlegroups.com
DELTA=108  (52 added, 36 deleted, 20 changed)
OCL=5932446
SCL=5933728