[Author: zork]
Initial API for adding event callbacks to localserver.  This allows a user to provide callbacks for error handling, changes in progress, and completion of an update.  The actual invocation of these callbacks will be added once we have a method for propagating the events from the update task to any associated ManagedResourceStore objects.

R=michaeln
CC=google-gears-eng@googlegroups.com
DELTA=416  (412 added, 2 deleted, 2 changed)
OCL=6022543
SCL=6112372