[Author: aa]
Implement JsCallContext for IE.

This changelist makes IE work similarly to Firefox -- IDL is used to
list what methods are available, but the methods in the IDL are declared
to take no parameters. Instead JsCallContext is used to get the params,
set return values, and exceptions.

For now, since we don't have our own dynamic dispatch implemented in
ModuleBaseClass, we have to use a ThreadLocal stack to store the
JsCallContext instances in so that we can get them later in the Gears code.
This will go away later and we will pass JsCallContext directly to the
Gears module.

R=cprince
CC=google-gears-eng@googlegroups.com
DELTA=687  (563 added, 102 deleted, 22 changed)
OCL=6007031
SCL=6139360