[Author: oshlack]
Implement type coercion for strings in Firefox and IE.
This changelist allows JsTokenToString to accept a JsToken
of any type, which will then be coerced to a string if it is not
already. It is backwards compatible with existing code that uses
js_types.h, the only difference is that JS code can now pass
non-string parameters to Gears where a string parameter is expected
and not have it throw an exception.

PRESUBMIT=passed
R=aa
CC=google-gears-eng@googlegroups.com
APPROVED=aa
DELTA=41  (35 added, 2 deleted, 4 changed)
OCL=6278975
SCL=6289229