[Author: aa]
Automated g4 rollback of changelist 6289229.

*** Reason for rollback ***

This change broke a bunch of the unit tests because some modules were not
expecting JsTokenToString() to coerce things that were not strings.

*** Original change description ***

Implement type coercion for strings in Firefox and IE.
This changelist allows JsTokenToString to accept a JsToken
of any type, which will then be coerced to a string if it is not
already. It is backwards compatible with existing code that uses
js_types.h, the only difference is that JS code can now pass
non-string parameters to Gears where a string parameter is expected
and not have it throw an exception.

PRESUBMIT=passed
TBR=oshlack
OCL=6300127
SCL=6300242