[Author: aa]
NOTE: I sent this CL out for review once before, but accidentally deleted it
before it got reviewed.

===

Rollback part of CL 6139360. That change implemented ModuleWrapper and 
JsCallContext for IE and made the GearsTest module use it. 

The implementation of ModuleWrapper is incomplete though and getting in the 
way of people using GearsTest for other things. 

This change leaves ModuleWrapper and JsCallContext checked in, but not 
hooked up to anything. It also reverts GearsTest to use COM and XPCOM.

PRESUBMIT=passed
R=marli
CC=google-gears-eng@googlegroups.com
APPROVED=oshlack
DELTA=121  (13 added, 98 deleted, 10 changed)
OCL=6342431
SCL=6344007