[Author: andreip]
Adds a setup dll to Gears installer on Wince. This dll exports functions that are called during the installation process. The purpose of the dll is to prompt the user to restart the device in order to complete the Gears installation. We have chosen to restart the device because requiring users to restart  just IE  is too tricky on Windows Mobile. We will remove this solution if we ever figure out a way to make IE reload BHOs at runtime (so that our extensions to the Tools menu would be visible without having to restart the browser).

- Edits to config.mk
  - Adding another DLL to the build system required that we move the program database (PDB) file name away from the linker flags variable.
 - We needed a separate "def" file for the new DLL.
-  We added /n flag to the resource compiler, to make it append null's to all strings in the string tables.

- Edits to rules.mk:
 - Added rules for building setup.dll.

R=michaeln,steveblock
CC=google-gears-eng@googlegroups.com
DELTA=271  (263 added, 0 deleted, 8 changed)
OCL=6286735
SCL=6367275