[Author: andreip]
Due to popular demand (e.g. QA), this CL changes slightly the install flow of Gears on WINCE. If we now fail to stop IE, we offer the choice to reboot the device. This is because on some Windows Mobile 6 standard devices, the process restarter fails to get a handle to the IE process (ErrorAccessDenied) and is therefore unable to kill it.

R=steveblock,cprince
CC=google-gears-eng@googlegroups.com
DELTA=106  (50 added, 37 deleted, 19 changed)
OCL=6497177
SCL=6499400