[Author: aa]
Automated g4 rollback of changelist 6483486.

*** Reason for rollback ***

This is a partial rollback of the original CL. Only the deletion of the
sdk/tools directory was rolled back. The addition of new files to the
inspector/ directory was not rolled back.

Rolling back the deletion until we complete integration of the inspector
and update the documentation to not refer to these files.

*** Original change description ***

Initial checkin of developer tools including log viewer tool, common
UI framework and support scripts. Also moved database and localserver
tools to gears/inspector/ and integrated into framework. Note that
these tools lack polish.

PRESUBMIT=passed
R=cprince
CC=gears-internal
DELTA=920  (920 added, 0 deleted, 0 changed)
OCL=6584593
SCL=6584710