[Author: steveblock]
Fixes dialog button text on WinCE.

The recent rearrangement of the TRANS_BLOCKs in the dialog HTML caused extra whitespace in the button labels on WinCE. This is fixed by adding a String.trim method in html_dialog.js, which is used to trim the innerText string before setting the button label.

The implementation of String.trim() uses a '^' character in a regex expression. The '^' character was also used as the m4 delimiter in the HTML files. The changes to the HTML files simply switch the m4 delimiter to '~'.

R=aa
CC=google-gears-eng@googlegroups.com
APPROVED=aa
DELTA=89  (7 added, 0 deleted, 82 changed)
OCL=6651401
SCL=6663615