[Author: aa]
Disable coercion in JsParamFetcher, JsCallContext, JsObject, and JsArray.
This makes the behavior of Gears in FF very similar to how it was in 0.2,
except that I have cleaned up some error handling to make it report
type mismatches better at the same time.

I moved the coercion code to CoerceJsTokenTo*() functions which can be
called on an as-needed basis for APIs that require it. Console is one
example that we have today.

PRESUBMIT=passed
R=cprince
CC=google-gears-eng@googlegroups.com
DELTA=556  (372 added, 79 deleted, 105 changed)
OCL=6735179
SCL=6748839