[Author: nigeltao]
Worker messages can now be JSON objects (more or less), not just
strings.  For more details, see the documentation at the beginning
of the js_pickle.h file.

This is bug 184:
http://code.google.com/p/google-gears/issues/detail?id=184

When receiving messages, i.e. implementing onmessage(a, b, m),
m.text is deprecated (although still functional), and you should
use m.contents instead (since the message's contents are not
necessarily textual, but may be numeric, an array, etc.).

This CL does not directly facilitate sending Blobs between
Workers, but is an intermediate step towards that (and will
presumably be easier once Blobs are Dispatcher-based).

Message (in workerpool.cc) was renamed to WorkerPoolMessage in
order to disambiguate it from base/common/message_*, which is a
different concept.

The SDK documentation will be updated in a future CL, once we
agree on the API design I have proposed in this CL.

R=aa
CC=google-gears-eng@googlegroups.com
APPROVED=aa
DELTA=1351  (1199 added, 63 deleted, 89 changed)
OCL=6684158
SCL=6832068