[Author: zork]
Non-Gecko-1.9-SDK portions of port of Google Gears for Firefox 3. Many thanks to Kimmo Kinnunen of Nokia for this ongoing work.
This includes a change to the extension so that the DLL will not register itself with Firefox if it's the wrong version.

NOTE This change does not include the Makefile changes required to make Firefox 2 and Firefox 3 versions of Gears co-exist in the same .xpi.

NOTE that this code is dormant; it compiles only when FF_MOD=experimental_19 is defined on the build command line.

R=cprince
CC=google-gears-eng@googlegroups.com
DELTA=656  (395 added, 162 deleted, 99 changed)
OCL=6607422
SCL=6838410