[Author: aa]
Fixes: http://code.google.com/p/google-gears/issues/detail?id=448

In FF3, width="100%" for things inside an absolutely positioned element no
longer make the parent expand. This is correct, according to the spec, but
not tradition. Anyway, forced the parent to be width 100% which required
moving it's padding down a level.

PRESUBMIT=passed
R=zork
CC=google-gears-eng@googlegroups.com
DELTA=3  (2 added, 1 deleted, 0 changed)
OCL=6852103
SCL=6856904