[Author: steveblock]
Modifies DocumentJsRunner::InvokeCallback and Eval for WinCE so that window.onerror is explicitly called if a script throws an exception.

Also modifies ThrowGlobalError for WinCE to ensure that window.onerror is only called in the context of the main page, not in a worker.

R=andreip,cprince
CC=google-gears-eng@googlegroups.com
APPROVED=andreip
DELTA=210  (136 added, 33 deleted, 41 changed)
OCL=6853991
SCL=6889724