[Author: aa]
This change updates JsonCpp to add escaping for strings. The code is taken
from the JsonCpp project itself, which has been fixed since we imported it.

Long term, it would be better to import an up-to-date copy of JsonCpp, but
I do not want to take on that risk right now.

PRESUBMIT=passed
R=cprince
CC=google-gears-eng@googlegroups.com
DELTA=82  (80 added, 0 deleted, 2 changed)
OCL=7038147
SCL=7052728