[Author: shess]
Manually patch in the following SQLite fixes:

http://www.sqlite.org/cvstrac/chngview?cn=4151
http://www.sqlite.org/cvstrac/chngview?cn=4183
This fixes a reentrancy problem with the fts virtual table in
case of schema change.

http://www.sqlite.org/cvstrac/chngview?cn=4095
http://www.sqlite.org/cvstrac/chngview?cn=4181
Fix to fts snippet-generation bug with leftmost column.

Unfortunately, the test files are likely to conflict on the
next vendor import, but that can be survived.

PRESUBMIT=passed
R=michaeln
CC=google-gears-eng
APPROVED=michaeln
DELTA=309  (248 added, 32 deleted, 29 changed)
OCL=7045328
SCL=7078819