[Author: aa]
Fix for issue 494: [FF] Running unit tests gives JScript error in
workerpool_message_body_tests.js.

Weird timing issue with BFCache in Firefox. This is not a great solution, but
I can't think of a better one at the moment. This bug was unrelated to 0.3
featureset, but more of a regression (an old one) in the test runner.

PRESUBMIT=passed
R=steveblock
CC=google-gears-eng@googlegroups.com
DELTA=8  (8 added, 0 deleted, 0 changed)
OCL=7222190
SCL=7225293