[Author: jripley]
Refactor Android ThreadMessageQueue to support both
asynchronous and blocking message loops on the main thread.

Every thread now has the same message queue structure. The
difference is that when a message is sent to the main thread,
the browser's message queue is also woken using
NPN_PluginThreadAsyncCall. This means that the main thread
queue will be serviced even if the main thread is in the
browser's idle loop. It also allows the queue to be run as a
manual, blocking loop when necessary.

To implement manually iterating the message queue, as is done
for the case of Win32 (GetMessage, Translate, Dispatch), there
is a new AndroidMessageLoop::RunOnce() method. This blocks
until at least one message has been dispatched.

This change is required to fix a problem with HttpRequest on
Android in the synchronous case, and to refactor
async_task_np.cc to be cross-platform.

PRESUBMIT=passed
R=andreip
CC=gears-eng@googlegroups.com
APPROVED=andreip
DELTA=156  (69 added, 28 deleted, 59 changed)
OCL=7414269
SCL=7420785