[Author: kart]
Upgrading to a newer version of Skia that supports color transforms. This CL deletes the existing version, and another CL will add the new version. This is because there are some 250 files to remove and an equivalent number to add, with substantial overlap, so to do it in one CL, I'll have to determine which files are in both versions and run g4 edit on them, determine the files that exist only in the new version and run g4 add on them, and g4 delete on the rest. This is a slow, manual and error-prone process when some 500 files are involved. Cleaner to wipe out the old version in one shot in this CL and send in another CL to add the new version. This will not break the build since nothing depends on these files yet.

PRESUBMIT=passed
R=nigeltao
CC=cprince,satish,aprasath
APPROVED=nigeltao
OCL=7474600
SCL=7480680