[Author: playmobil]
* Fix so that a binary compiled on 10.5 runs correctly on 10.4.
* Fix a couple of small issues leftover form the Google Gears->Gears name change for Safari.
* Add clean_gears.sh script to uninstall gears for Safari.

The problem with 10.5 & 10.4 has to do with an Apple bug which can be
read about in more detail at	http://developer.apple.com/technotes/tn2007/tn2185.html#TNTAG18

We compile Gears for Safari with all symbols hidden by default, only exporting the
necessary NPAPI entry points.
To workaround the aformentioned bug we need to declare the C++ <string> symbols as visible.

The problem with the way we handled this previously, was that while string16.h had
the correct compiler flags to make things work correctly, we still have source files
that include <string> but not string16.h, this confused the linker under 10.5 as
some files had <string> marked as being hidden (those that included <string> but not string16),
and some had it marked as visible (those including string16.h).
In addition, the underlying bug has apparently been fixed in 10.5, which made
me think that hiding <string> when compiling on Leoard would solve the
issue, but instead it only made binaries compiled on 10.5 not work on 10.4.

This CL hopefully fixes this issue by adding a pre-include file to all C++ files compiled
for Safari that adds the <string> header with it's symbols marked as visible.  This makes
the linker happy, and works around the bug so that binaries compiled on 10.5 work on 10.4.

PRESUBMIT=passed
R=michaeln
CC=gears-internal
APPROVED=michaeln
DELTA=95  (67 added, 12 deleted, 16 changed)
OCL=7507281
SCL=7518901