[Author: kart]
Makefile for Skia and a few changes in Gears code to make it build. The resulting binaries are placed in third_party/skia/bin-OS-ARCH. This is a separate CL from the skia source itself because of a g4/mondrian bug.

PRESUBMIT=passed
R=cprince
CC=aa,satish,aprasath,kart
APPROVED=cprince
OCL=7443890
SCL=7520877