[Author: michaeln]
Fix a bug with redirection. In our manifest file syntax it is possible to express requests for a particular url to be redirected to one and the same url. The public docs service contains an entry that does so. This change detects those entries and ignores them when servicing urls from the localserver module.

R=playmobil
CC=gears-eng@googlegroups.com
DELTA=38  (14 added, 0 deleted, 24 changed)
OCL=7640289
SCL=7649025