[Author: nigeltao]
LocalServer and ResourceStore are now Dispatcher-based. In doing so,
ResourceStore::CaptureBlob is now implemented on NPAPI.

This also means that GearsFactory is the only non-Dispatcher module
remaining, and so, for example, no longer has a CreateComModule method.

The LocalServer merge was fairly trivial. ResourceStore merging was
less so, because of the platform-specific versions of AsyncTask have
different APIs. For example, IE's takes a HWND. Also, CaptureBlob was,
prior to this CL, not implemented on NPAPI.

There were also substantial differences in FF's and IE's implementaion
of ResourceStore::CaptureFile, but it was straightforward to re-write
CaptureFile to just create a FileBlob and then follow CaptureBlob's
implementation, avoiding the need for, e.g., resource_store.cc to have
its own, platform-specific "read this file into memory" code.

PRESUBMIT=passed
R=michaeln
CC=gears-eng@googlegroups.com
APPROVED=michaeln
DELTA=6542  (1755 added, 4771 deleted, 16 changed)
OCL=7587507
SCL=7655347