[Author: playmobil]
Temporary workaround for bug 568 to enable us to turn on Safari unit tests 
in pulse.

This change reads an NSInputStream into an in-memory buffer and sends that over
the wire rather than a stream.  This doesn't trigger the bug and posts go through
just fine.

This is however only a temporary workaround, we must fix this before launch, tracking
this in bug 568.

PRESUBMIT=passed
R=bgarcia
CC=gears-eng@googlegroups.com
DELTA=20  (19 added, 0 deleted, 1 changed)
OCL=7673117
SCL=7673420