[Author: kart]
Adding some unit tests, implementing Clone and Crop and fixing some bugs. In canvas_tests.js, I moved common functions to the top of the file and reordered the test cases to be in the same order as the methods in the .h and .cc files.

 Many of the tests use a golden file, which stores the expected output. We perform the image operation and compare the actual output with the expected output in the golden file. If they match, we're good. Some of the tests also perform checks on the golden file itself. If someone accidentally replaces the golden file, the test could fail and we would mistakenly conclude the code being tested has a bug. To avoid such cases, some tests also check the golden file.

  The following functions were added in canvas_tests.js: testClone() and its helper runCloneTest(), testCrop(), testCropToZeroSize(), testCropNoop(), testResizeWeirdCases(). The others are mostly not changed.

PRESUBMIT=passed
R=aa
CC=satish,kart,gears-internal,bgarcia
APPROVED=bgarcia,cprince
OCL=7655456
SCL=7691463