[Author: jripley]
Use MockAudioRecorder on Android for now.

We're lacking a PortAudio back-end for the audio subsystem on
Android, so for now it's sufficient to use the
MockAudioRecorder factory instead of PaAudioRecorder. This
satisfies linking and presents an interface which succeeds all
calls (and does nothing). This is only for Android.

PRESUBMIT=passed
R=vamsikrishna,andreip
CC=gears-eng@googlegroups.com
APPROVED=andreip
DELTA=9  (9 added, 0 deleted, 0 changed)
OCL=7692646
SCL=7697289