[Author: playmobil]
The browser name stored in the database was erroneously set to 'npapi' for the Safari
port.  This caused an error when trying to use a database created with an old version
which had the browser name set correctly.

This CL fixes the browser name and removes the check, which doesn't seem so important.

PRESUBMIT=passed
R=mpcomplete
CC=gears-eng@googlegroups.com
DELTA=19  (10 added, 7 deleted, 2 changed)
OCL=7792643
SCL=7792925