[Author: jripley]
Delete temporary files when closed on POSIX.

POSIX builds are currently leaking temporary files because
they are not deleted with the File instance, unlike Win32
builds. For example, HttpRequests of objects greater than the
ByteStore threshold (1MB on most platforms) will leak.

Also added a File::GetBaseTemporaryDirectory() so that
platforms can define a location for temporary files other than
the default "/tmp". On Android, this is a subdirectory of the
browser so that stale contents can be easily cleaned on
startup, and because /tmp is actually backed by ramdisk.

PRESUBMIT=passed
R=andreip,playmobil
CC=gears-eng@googlegroups.com
APPROVED=andreip
DELTA=280  (210 added, 46 deleted, 24 changed)
OCL=7773642
SCL=7796664