[Author: aa]
Refactor upgrade logic in PermissionsDB to reduce duplication.

Added a public utility method to SQLDatabase to find out whether it's in a
transaction.

Removed all the logging. I don't think it's necessary because the upgrade is
in a transaction. Whatever was causing an upgrade to fail should be
consistent, and just looking at the database should give better insight that
the log file.

PRESUBMIT=passed
R=michaeln
CC=gears-eng@googlegroups.com
DELTA=285  (48 added, 221 deleted, 16 changed)
OCL=7791785
SCL=7805403