[Author: michaeln]
Validate content prior to accepting responses for insertion into the localserverDB.

The problem:

When a network error occurs mid-stream during resource fetching, gears cannot reliably detect that something has gone wrong. One source of this problem is in firefox, the nsHttpChannel interface has a means to expose this type of error to its clients (gears), however it fails to do so. A consequence of this is that gears inserts partially received resposes into its cache (not good!!).

We think this problem is the cause of the partial captures being reported, and also the cause of oddly formed degenerate responses (no headers/no data) being passed up to gears on "successful" requests in firefox.

Additionally, gears depends on each hosting browser's http infrastructure as a transport layer. The specific problem we're aware of today applies to firefox (1.5, 2, and 3), its very difficult to know that our interactions with other browsers stacks don't also contain failure modes of this type.

What to do about it:

Optionally validate data received from the browser's stack prior to accepting it. This will help us avoid the specific problem we know of with firefox today, and protect us from the ones we don't yet know of.

Details:

A light weight flavor of validation is testing an expected content length to the amount of data actually recieved.

To that end, we define a custom http response header that contains the decoded content length.

   X-Gears-Decoded-Content-Length: 1234

  This header differs from the standard Content-Length header in a couple of ways.
   * describes the size of the decoded data, the std header describes the encoded data
   * can be applied to chunked responses w/o bumping heads with the http spec which
     indicates that chunked responses should NOT contain content length headers

  Also, its difficult to 'trust' many of the content-length headers out there. Nothing goes wrong if the value
  is inaccurate in many cases, so we are wary of testing against that std header value.

3) Validation steps

* Validation is optional

* If a response contains the X-Gears-Decoded-Content-Length header, the response will only be accepted into the cache if the received length matches. This step happens for all resources being added to the cache and may be used to validate ResourceStore.capture() response data as well as ManagedResourceStore manifest file entries.

R=cprince
CC=gears-internal
APPROVED=ace
DELTA=219  (147 added, 41 deleted, 31 changed)
OCL=7764010
SCL=7854832