[Author: nigeltao]
Uncrash Firefox's js_runner_ff_marshaling *on Windows*, introduced by
7845800 "Take the linked_ptr's out of js_runner_ff_marshaling".

The base problem was that JsWrapperDataForProto were being deleted (via
Spidermonkey's finalization) before the JsWrapperDataForInstances were.
However, the JsWrapperDataForProto contains a scoped_ptr to a JSClass,
and that JSClass is how Spidermonkey knows to call FinalizeNative on the
JsWrapperDataForInstance.

For Linux, things would almost always "work", but on Windows (and
specifically, Visual Studio's compiler), deleting the JSClass object
would fill the memory it used to live in with 0xDD values, and hence
JSClass::finalize would no longer point to the address of the
FinalizeNative function, and hence finalizing a JS object for a Gears
module would crash.

The solution is to remove the ownership (in the scoped_ptr sense) of the
JSClass from the JsWrapperDataForProto, and instead give it to the
JsContextWrapper, which already keeps track of a std::set of all
allocated JSClass* objects.

R=cprince
CC=gears-eng@googlegroups.com
APPROVED=cprince
DELTA=180  (77 added, 46 deleted, 57 changed)
OCL=7909299
SCL=7920608