[Author: playmobil]
In Safari, when a proxy is set from a PAC file the JS in that file will be executed
on it's own thread which can cause a gc.  If that gc tries to kill a Gears object, that
object will try to use a copy of the npn function table stored in Thread Local Storage,
however since the table doesn't exist on the new thread we crash.

This CL fixes this by forwarding delete calls to the main thread if the function table
isn't initialized.

PRESUBMIT=passed
R=mpcomplete
CC=gears-eng@googlegroups.com
DELTA=49  (48 added, 0 deleted, 1 changed)
OCL=8063245
SCL=8064019