[Author: nigeltao]
Make FF's DocumentJsRunner self-deletion occur later,
specifically at the later of (1) the page unload event
being fired, and (2) the ModuleEnvironment (that is
shared by all Gears modules for that DocumentJsRunner)
being destroyed. Prior to this CL, deletion was
happening at only (1), which leads to a crasher if
Gears modules are used during window.unload, such as
the code snippet given on issue 270.

JsRunnerBase::ForceGC on Firefox is once again guarded by
an #ifdef DEBUG, since it is no longer always needed -
the #ifdef DEBUG was removed in 7871060 because a GC was
used to attempt to ensure the DocumentJsRunner was deleted
after the ModuleEnvironment was deleted. That attempt was
not fully successful, as evidenced obviously by said
crasher.

PRESUBMIT=passed
R=aa
CC=gears-eng@googlegroups.com
APPROVED=michaeln
DELTA=70  (58 added, 8 deleted, 4 changed)
OCL=8040339
SCL=8066351