[Author: nigeltao]
Fix drag-and-drop Firefox crasher. Previously, we were explicitly
deleting the DropTarget object. This was a latent bug that wasn't
problematic when the original feature was submitted, presumably
because we were leaking a number of things on Firefox, at the time.

Now that the leaks have been fixed, the explicit deletion of the
XPCOM (i.e. ref-counted) object is causing a crash when Firefox's
EventListener management code tries to follow a dangling pointer.
Instead, the correct thing to do (which is what this patch
introduces) is to balance any existing AddEventListener (and hence
AddRef) calls with RemoveEventListener (and hence Release) calls.

PRESUBMIT=passed
R=noel
CC=gears-eng@googlegroups.com
APPROVED=noel
DELTA=103  (72 added, 24 deleted, 7 changed)
OCL=8573993
SCL=8574080