[Author: steveblock]
Automated g4 rollback of changelist 8626593.

Removes unused strings from permissions_dialog.js.stab. The strings in the TC are now fixed, so the translated strings can now be updated.

*** Original change description ***

Automated g4 rollback of changelist 8618774.

This is a partial rollback. It rolls back the deletion of unused strings from permissions_dialog.js.stab, because there are currently problems with the latest strings from the translation console, so the translated strings can not be updated to reflect the deletion.

The strings will be removed and the translated strings updated once correct translations are available.

R=andreip
CC=gears-internal
DELTA=22  (0 added, 22 deleted, 0 changed)
OCL=8744852
SCL=8745405