[Author: jripley]
Refactor HttpRequestAndroid to use SafeHttpRequest.

The Android browser had a late change before launch which
required the cache to only be accessed by the main
thread. This complicated the state machine and created a
deadlock in POST shutdown.

This change refactors HttpRequestAndroid to use
SafeHttpRequest instead, so it is only ever involves two
threads: the main and child thread, and not a third (the
worker or async task).

This also changes inter-thread hand-off to an explicit state
machine, which has a much clearer execution path. The state
machine is only ever incremented by the main thread, and very
little the class state is updated by the child, now.

There is an extra method
AndroidMessageLoop::RunOnceWithTimeout(), which allows the
message loop to the iterated with a timeout per call. This is
to implement the exponential backoff based
java.net.URLConnection abort.

PRESUBMIT=passed
BUG=1416046
R=andreip,nicolasroard
CC=gears-internal
DELTA=972  (478 added, 276 deleted, 218 changed)
OCL=8606673
SCL=8971876