[Author: jripley]
Move the bulk of Android Gears to opensource.

This moves all files that do not require hand-merging. In most
cases the path "private_android" is simply replaced with
"opensource". There are a few exceptions:

base/common/paths_np.cc was renamed to paths_android.cc, and
the opensource Makefile changed to reflect this. It was not
merged with paths_np.cc as the implementation is totally
different. paths_np.cc needs renaming to paths_win32.cc.

base/npapi/module.cc was renamed to module_android.cc, and the
opensource Makefile changed to reflect this. The
implementations are similar but have enough subtle differences
to warrant a fork, or at least a future TODO to refactor them.

The Makefiles and tools have not been moved as they need
hand-merging. The installer has not been moved as it is now
cruft (but potentially useful in future). The Makefiles in
private_android have been modified to reference the new paths
instead.

PRESUBMIT=passed
R=andreip,cprince
CC=gears-internal
OCL=8974124
SCL=9007535