[Author: jripley]
Fix comparison of Dispatcher::ImplCallback with NULL.

Dispatcher::ImplCallback is a templated pointer to member
function. On some platforms (e.g Android, which is ARM EABI),
these pointers are actually two words (64 bits), where the
second word indicates whether the first is a plain pointer
(static or non-virtual member), or an offset in the vtable
(virtual member).

The problem is that the first virtual function in a vtable, at
offset 0, is encoded as: 0x00000000, 0x00000001. Implicit cast
to bool results in false, and it compares as equal to NULL
(casting NULL to a pointer-to-vtable-offset results in the
same encoding). This means that RegisterProperty() fails if
the getter is a virtual function at offset 0.

It also means that SetProperty() will fail if the setter
function is at virtual offset 0, as NULL is used as a special
value to indicate read-only. (This does not trigger at the
moment, by luck).

This fix involves using a union to check the raw binary value
of the pointer. This does still assume that NULL will map to
all-zero, and vtable offset 0 will not, but this seems safe on
current platforms.

PRESUBMIT=passed
R=andreip,nigeltao
CC=gears-eng@googlegroups.com
DELTA=45  (43 added, 0 deleted, 2 changed)
OCL=8972115
SCL=9031462