[Author: nigeltao]
Introduce a this_arg parameter to JsRunnerInterface::InvokeCallback.

This is part 1 of a three-part ChangeList.

Part 1 (this part) does not change any functionality, it simply
adds an argument (which is always NULL, at this stage) to a method.
This CL is basically nothing but boilerplate code changes, separated
out into a stand-alone CL so that the two remaining CLs are not
cluttered by dozens of lines of boilerplate changes, so that the
reviewer can more easily focus on the deltas that matter.

Part 2 will actually implement the this_arg feature (and add
the automated test cases). For example, in js_runner_ff.cc,
this CL will modify InvokeCallbackSpecialized to not assume
JS_CallFunctionValue should take the result of JS_GetGlobalObject.

Part 3 will update desktop/drop_target_sf.cc to actually pass a
non-NULL this_arg, since currently it has to workaround the "always
assume this == theGlobalObject" thing when it wants to call
event.preventDefault(), since preventDefault assumes that "this ==
event", and hence will fail with a TypeError if we try to
JsRunnerInterface::InvokeCallback it.

R=aa
CC=gears-eng@googlegroups.com
APPROVED=mpcomplete
DELTA=62  (21 added, 0 deleted, 41 changed)
OCL=9229225
SCL=9260656