[Author: jripley]
Deny fts{2,3}_tokenizer usage with Android SQLite.

The fts3_tokenizer is enabled in the system Android SQLite,
which potentially allows arbitrary blob pointers to be passed
directly from scripts.

Rather than modifying the system SQLite, which will be pushed
relatively slowly, this change adds a check in the
ForbidActions authorizer. When the tokenizer is used, it is
called back with SQLITE_FUNCTION and argument
"fts3_tokenizer". If ForbidActions is called with this value
(or fts2_tokenizer), it is denied.

PRESUBMIT=passed
BUG=1384455
R=andreip,shess
CC=gears-internal
DELTA=8  (8 added, 0 deleted, 0 changed)
OCL=9368241
SCL=9387522