[Author: jripley]
Fix database unit tests for different Android errors.

The system SQLite on Android has the load_extension and
fts3_tokenizer functions compiled in, but disabled. This means
that the error messages thrown by using them are not "no such
function", and are instead "logic error" and "authorization
denied", respectively.

This also disables testing the exact value of the error
message in a worker, as Android appears to be suffering from
the same SpiderMonkey related problem as Safari - the error
message text is not propagating.

PRESUBMIT=passed
BUG=1384455
R=andreip,zakcohen
CC=gears-internal
APPROVED=zakcohen
DELTA=40  (30 added, 0 deleted, 10 changed)
OCL=9368646
SCL=9387597