[Author: jripley]
Add progress messages to HttpRequestAndroid.

This implements both download and upload progress messages for
HttpRequests on Android.

Due to the transfers actually occurring on a background
thread, this requires that the message is actually performed
by the main thread via a ProgressEvent, which uses an
AsyncFunctor. HttpRequestAndroid implements the
ProgressEvent::Listener interface.

If the HttpRequest was created by a worker thread, this is
forwarded again to the worker via the wrapper in
SafeHttpRequest.

PRESUBMIT=passed
BUG=1492015
R=andreip,nicolasroard
CC=gears-eng@googlegroups.com
DELTA=59  (54 added, 4 deleted, 1 changed)
OCL=9387431
SCL=9449691