[Author: michaeln]
Use BINDSTATUS_RAWMIMETYPE when reporting the mimetype of cached resources. Otherwise IE's image loading can get completely wedged if a GIF is reported as a PNG when using BINDSTATUS_MIMETYPEAVAILABLE, wedged to the point of no images will load for the duration of the process. This is only an issue when using HttpHandlerPatch.

Tested on IE7/XPSP2.

R=andreip
CC=gears-internal
DELTA=21  (13 added, 7 deleted, 1 changed)
OCL=10115977
SCL=10127351